Localize RM-18 Ollama token counting
